Headers not sent when pushing streams
When pushing a stream of a `<script type="module">/* ... */</script>` to the browser, the runtime would error out with:

```
Failed to load module script: The server responded with a non-JavaScript MIME type of "". Strict MIME type checking is enforced for module scripts per HTML spec.
```
It came to my attention that the headers were missing (only `status` was present). The changes here address that.